requirements.txt
asgiref==3.8.1
boto3==1.26.17
botocore==1.29.165
Django==3.2
django-cors-headers==3.13.0
django-storages==1.11.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.7.0
gunicorn==23.0.0
jmespath==1.0.1
packaging==24.1
Pillow==10.0.0
PyJWT==2.9.0
pymongo==3.12.0
python-dateutil==2.9.0.post0
pytz==2024.2
s3transfer==0.6.2
setuptools==75.2.0
six==1.16.0
sqlparse==0.2.4
urllib3==1.26.20